feat: Bootstrap the project using standalone APIs
