+++ src/app/pages/portal/toast-panel/toast-panel.component.ts
@@ -77,19 +77,19 @@ interface ExtendedGeoServiceDefinition extends GeoServiceDefinition {
   changeDetection: ChangeDetectionStrategy.OnPush,
   standalone: true,
   imports: [
-    NgIf,
-    PanelComponent,
-    NgClass,
     ActionbarComponent,
-    MatButtonModule,
-    StopPropagationDirective,
-    MatTooltipModule,
-    MatIconModule,
-    MatBadgeModule,
-    FeatureDetailsComponent,
-    SearchResultsComponent,
-    NgTemplateOutlet,
     AsyncPipe,
+    FeatureDetailsComponent,
+    MatBadgeModule,
+    MatButtonModule,
+    MatIconModule,
+    MatTooltipModule,
+    NgClass,
+    NgIf,
+    NgTemplateOutlet,
+    PanelComponent,
+    SearchResultsComponent,
+    StopPropagationDirective,
     TranslateModule
   ]
 })
